更正网络结构规律读取

+++ src/cn/InstFS/wkr/NetworkMining/Miner/ProtocolAssociationLine.java
@@ -276,7 +276,8 @@ public class ProtocolAssociationLine {
 					int trafficSum  = 0;
 					for(int k = j;k < size;k++)
 					{
-						trafficSum += Integer.parseInt(pdi.dataItems.data.get(k));
+//						trafficSum += Integer.parseInt(pdi.dataItems.data.get(k));
+						trafficSum += Double.parseDouble(pdi.dataItems.data.get(k));
 					}
 					newData.add1Data(pdi.dataItems.time.get(j), trafficSum+"");
 				}
